+++ src/main/java/com/example/demo/client/base/BaseClient.java
@@ -1,4 +1,4 @@
-package com.example.demo.client.baseClient;
+package com.example.demo.client.base;
 
 import org.springframework.http.HttpHeaders;
 import org.springframework.util.MultiValueMap;
